Added in ability to control delay (if wanted) for particle tracks within the detector.
This method worked much better then adding more points to the array and saved lots of memory so it will be implimented. Reverting this change only required a 10 to be converted to a 0 within the common functions file. This change seems to fix the problem of the particle tracks moving too fast within the detector so it was implimented.